Refactorizar server en app + index
